+++ Client/src/Models/Variables/IVariableNameModel.ts
